+++ gufe/storage/errors.py
@@ -10,5 +10,3 @@ class MissingExternalResourceError(ExternalResourceError):
 
 class ChangedExternalResourceError(ExternalResourceError):
     """Error when there's a SHA256 mismatch with an external resource"""
-
-
